+++ CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
 - Fix secretsProvider not being applied to new stacks [#935](https://github.com/pulumi/pulumi-kubernetes-operator/issues/935)
 - Fix Helm template YAML parsing error for podLabels [#1014](https://github.com/pulumi/pulumi-kubernetes-operator/issues/1014)
 - Fix Stack deletion being blocked by missing prerequisites [#751](https://github.com/pulumi/pulumi-kubernetes-operator/issues/751)
+- Fix TTL not being respected for completed Updates [#960](https://github.com/pulumi/pulumi-kubernetes-operator/issues/960)
 
 ## 2.2.0 (2025-08-11)